requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
python-multipart==0.0.6
google-genai==0.6.0
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-oauth2-credentials==0.1.0
llama-index==0.12.46